Creating background helper
